requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coverage==3.7.1
coveralls==1.1
cryptography==2.6.1
decorator==4.0.11
docopt==0.6.2
filetype==1.0.5
future==0.17.1
idna==2.8
imageio==2.1.2
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
instapy-cli==0.0.12
more-itertools==7.0.0
moviepy==0.2.3.2
numpy==1.16.3
Pillow==6.0.0
pluggy==0.11.0
py==1.8.0
pycparser==2.19
pytest==4.5.0
pytest-cov==2.5.1
python-telegram-bot==11.1.0
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.3
TexSoup==0.2.0
tqdm==4.11.2
urllib3==1.25.2
wcwidth==0.1.7
